requirements.txt
appdirs==1.4.4
attrs==20.2.0
aws-cdk.assets==1.73.0
aws-cdk.aws-amplify==1.73.0
aws-cdk.aws-apigateway==1.73.0
aws-cdk.aws-apigatewayv2==1.73.0
aws-cdk.aws-apigatewayv2-integrations==1.73.0
aws-cdk.aws-applicationautoscaling==1.73.0
aws-cdk.aws-autoscaling==1.73.0
aws-cdk.aws-autoscaling-common==1.73.0
aws-cdk.aws-autoscaling-hooktargets==1.73.0
aws-cdk.aws-batch==1.73.0
aws-cdk.aws-certificatemanager==1.73.0
aws-cdk.aws-cloudformation==1.73.0
aws-cdk.aws-cloudfront==1.73.0
aws-cdk.aws-cloudwatch==1.73.0
aws-cdk.aws-codebuild==1.73.0
aws-cdk.aws-codecommit==1.73.0
aws-cdk.aws-codeguruprofiler==1.73.0
aws-cdk.aws-codepipeline==1.73.0
aws-cdk.aws-cognito==1.73.0
aws-cdk.aws-dynamodb==1.73.0
aws-cdk.aws-ec2==1.73.0
aws-cdk.aws-ecr==1.73.0
aws-cdk.aws-ecr-assets==1.73.0
aws-cdk.aws-ecs==1.73.0
aws-cdk.aws-ecs-patterns==1.73.0
aws-cdk.aws-efs==1.73.0
aws-cdk.aws-elasticloadbalancing==1.73.0
aws-cdk.aws-elasticloadbalancingv2==1.73.0
aws-cdk.aws-events==1.73.0
aws-cdk.aws-events-targets==1.73.0
aws-cdk.aws-iam==1.73.0
aws-cdk.aws-kinesis==1.73.0
aws-cdk.aws-kinesisfirehose==1.73.0
aws-cdk.aws-kms==1.73.0
aws-cdk.aws-lambda==1.73.0
aws-cdk.aws-lambda-python==1.73.0
aws-cdk.aws-lambda-event-sources==1.73.0
aws-cdk.aws-logs==1.73.0
aws-cdk.aws-route53==1.73.0
aws-cdk.aws-route53-targets==1.73.0
aws-cdk.aws-s3==1.73.0
aws-cdk.aws-s3-assets==1.73.0
aws-cdk.aws-sam==1.73.0
aws-cdk.aws-secretsmanager==1.73.0
aws-cdk.aws-servicediscovery==1.73.0
aws-cdk.aws-sns==1.73.0
aws-cdk.aws-sns-subscriptions==1.73.0
aws-cdk.aws-sqs==1.73.0
aws-cdk.aws-ssm==1.73.0
aws-cdk.aws-stepfunctions==1.73.0
aws-cdk.cdk==0.36.1
aws-cdk.cloud-assembly-schema==1.73.0
aws-cdk.core==1.73.0
aws-cdk.custom-resources==1.73.0
aws-cdk.cx-api==1.73.0
aws-cdk.region-info==1.73.0
boto3==1.16.9
boto3-stubs==1.16.13.0
botocore==1.19.9
CacheControl==0.12.6
cachy==0.3.0
cattrs==1.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cleo==0.8.1
clikit==0.6.2
constructs==3.2.9
crashtest==0.3.1
cryptography==3.2.1
distlib==0.3.1
filelock==3.0.12
GoDaddyPy==2.3.3
html5lib==1.1
idna==2.10
iniconfig==1.1.1
jeepney==0.4.3
jmespath==0.10.0
jsii==1.14.0
keyring==21.4.0
lockfile==0.12.2
msgpack==1.0.0
packaging==20.4
pastel==0.2.1
pexpect==4.8.0
pkginfo==1.6.1
pluggy==0.13.1
poetry-core==1.0.0
ptyprocess==0.6.0
publication==0.0.3
py==1.9.0
pycparser==2.20
pylev==1.3.0
pyparsing==2.4.7
pytest==6.1.2
python-dateutil==2.8.1
python-terraform==0.10.1
requests==2.24.0
requests-toolbelt==0.9.1
s3transfer==0.3.3
SecretStorage==3.1.2
shellingham==1.3.2
six==1.15.0
toml==0.10.2
tomlkit==0.7.0
typing-extensions==3.7.4.3
urllib3==1.25.11
virtualenv==20.1.0
webencodings==0.5.1

Flask~=1.1.2
setuptools~=50.3.0.post20201006